+++ frontend/src/hooks/api/getters/usePendingChangeRequests/usePendingChangeRequests.ts
@@ -1,7 +1,12 @@
 import { formatApiPath } from 'utils/formatPath';
 import handleErrorResponses from '../httpErrorResponseHandler.js';
-import type { ChangeRequestType } from 'component/changeRequest/changeRequest.types';
+import type {
+    ChangeRequestType,
+    IFeatureChange,
+} from 'component/changeRequest/changeRequest.types';
 import { useEnterpriseSWR } from '../useEnterpriseSWR/useEnterpriseSWR.js';
+import { useMemo } from 'react';
+import { addConstraintIdsToFeatureChange } from 'utils/addConstraintIdsToFeatureChange.js';
 
 const fetcher = (path: string) => {
     return fetch(path)
@@ -16,8 +21,29 @@ export const usePendingChangeRequests = (project: string) => {
         fetcher,
     );
 
+    const mappedData: typeof data = useMemo(
+        () =>
+            data?.map((changeRequest) => {
+                const { features, ...rest } = changeRequest || {};
+                const featuresWithConstraintIds =
+                    features?.map((feature) => {
+                        const changes: IFeatureChange[] = feature.changes.map(
+                            addConstraintIdsToFeatureChange,
+                        );
+
+                        return {
+                            ...feature,
+                            changes,
+                        };
+                    }) ?? [];
+
+                return { ...rest, features: featuresWithConstraintIds };
+            }),
+        [JSON.stringify(data)],
+    );
+
     return {
-        data,
+        mappedData,
         loading: !error && !data,
         refetch: mutate,
         error,
